Use jinja2 templates directly instead of relying on `PromptTemplate`
Now `ChatSchema` stores directly a jinja2 template instead of a
langchain `ChatPromptTemplate` with jinja2 support enabled. This allows
for better filter customization to support complex model-specific prompt
templates.

Also, the tool-calling prompt templates (llama3.1 and mistral large)
have been modified to use the new custom filters to detect the last user
message in the list of message. This is required in order to format the
prompt with available tools in the right order (just before last user
message in mistral, inside last user message in llama3.1).

Also added a new `to_json` custom filter, used to properly format the
list of arguments inside `AIMessages` with tool_calls. Without this,
jinja2 formats the arguments as a plain python dict. This risks confusing
the models on the right quotes to use for tool call generation (' vs ").